Added some Validators
