+++ .gitignore
@@ -4,6 +4,9 @@ target/
 !**/src/main/**
 !**/src/test/**
 
+#Logs
+.log
+
 ### STS ###
 .apt_generated
 .classpath
