--- src/main/java/com/noted/api/api/DTO/NoteDTO.java
@@ -1,8 +1,0 @@
-package com.noted.api.api.DTO;
-
-public class NoteDTO {
-    public String title;
-    public String text;
-    public String notebook_id;
-}
-
