--- src/main/java/com/noted/api/api/DTO/NotebookDTO.java
@@ -1,6 +1,0 @@
-package com.noted.api.api.DTO;
-
-public class NotebookDTO {
-    public String name;
-    public String id;
-}
