+++ src/main/java/com/noted/api/model/Notebook.java
@@ -5,6 +5,7 @@ import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.Id;
 import javax.persistence.OneToMany;
+import javax.validation.constraints.NotNull;
 
 import java.util.List;
 import java.util.UUID;
@@ -13,7 +14,9 @@ import java.util.UUID;
 public class Notebook {
 
     @Id
+    @NotNull(message = "A Notebook must be created with an ID")
     private UUID id;
+    @NotNull(message = "A Notebook must contain a name")
     private String name;
 
     @OneToMany(fetch = FetchType.LAZY, mappedBy = "notebook", cascade = CascadeType.ALL)
